[iOS] Change firstResponderPublisher from PassthroughSubject to CurrentValueSubject
This makes it so the first call will return a responder if the firstResponder was set before we started listening.

This fixes an issue where the firstResponder was set before the ResponderChain was initialised.

This also makes it easier to test the windows firstResponder cross platform because the macOS firstResponderPublisher works the same.